ocr-service/requirements-test.txt
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
pillow>=10.0.0
numpy>=1.26.0
opencv-python-headless>=4.10.0
